tec: Move CurrentUser under flusio\auth\ namespace
