+++ src/controllers/collections/Bookmarks.php
@@ -3,8 +3,8 @@
 namespace flusio\controllers\collections;
 
 use Minz\Response;
+use flusio\auth;
 use flusio\models;
-use flusio\utils;
 
 /**
  * @author  Marien Fressinaud <dev@marienfressinaud.fr>
@@ -21,7 +21,7 @@ class Bookmarks
      */
     public function show()
     {
-        $user = utils\CurrentUser::get();
+        $user = auth\CurrentUser::get();
         if (!$user) {
             return Response::redirect('login', [
                 'redirect_to' => \Minz\Url::for('bookmarks'),
