+++ src/controllers/collections/Discovery.php
@@ -3,6 +3,7 @@
 namespace flusio\controllers\collections;
 
 use Minz\Response;
+use flusio\auth;
 use flusio\models;
 use flusio\utils;
 
@@ -23,7 +24,7 @@ class Discovery
      */
     public function show($request)
     {
-        $user = utils\CurrentUser::get();
+        $user = auth\CurrentUser::get();
         if (!$user) {
             return Response::redirect('login', [
                 'redirect_to' => \Minz\Url::for('discover collections'),
